YEL-68 [refactor] Util 클래스 의미가 모호하여 Factory 객체로 정정
