+++ src/main/java/com/yello/server/global/common/factory/PaginationFactory.java
@@ -1,4 +1,4 @@
-package com.yello.server.global.common.util;
+package com.yello.server.global.common.factory;
 
 import org.springframework.data.domain.*;
